+++ src/cutsel.c
@@ -134,7 +134,7 @@ static void LengthReceived(Widget w, XtPointer client_data, Atom *selection,
   if (*type == 0)
     printf("No length received\n");
   else if (*type == XA_INTEGER) {
-      printf("Length is %" PRIx32 "\n", *(CARD32*)value);
+      printf("Length is %" PRIx32 "\n", (uint32_t)*(CARD32*)value);
   } else {
       char *name = XGetAtomName(d, *type);
       printf("Invalid type received: %s\n", name);
